libs/ecs/src/resource.ts
import {
  AbsentFilter,
  AddedFilter,
  ChangedFilter,
  PresentFilter,
} from "./filter";

/**
 * A base class for Resources.
 */
export default abstract class Resource {
  /**
   * Marker to identify resoruces.
   *
   * @returns true
   */
  public __isResource(): boolean {
    return true;
  }

  /**
   * Create a filter that only allows containers that have the trait.
   *
   * @returns An instance of PresentFilter
   */
  public static present(): PresentFilter<Resource> {
    return new PresentFilter(this);
  }

  /**
   * Create a filter that only allows containers that do not have the trait.
   *
   * @returns An instance of AbsentFilter
   */
  public static absent(): AbsentFilter<Resource> {
    return new AbsentFilter(this);
  }

  /**
   * Create a filter that only allows containers that gained trait
   * since last query reset.
   *
   * @returns An instance of AddedFilter
   */
  public static added(): AddedFilter<Resource> {
    return new AddedFilter(this);
  }

  /**
   * Create a filter that only allows containers that had the trait
   * changed since last query reset.
   *
   * @returns An instance of ChangedFilter
   */
  public static changed(): ChangedFilter<Resource> {
    return new ChangedFilter(this);
  }
}
